+++ BitmovinAnalyticsCollector/Classes/BitmovinPlayer/Classes/BitmovinPlayerAdapter.swift
@@ -4,12 +4,14 @@ import BitmovinPlayer
 class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
     private let config: BitmovinAnalyticsConfig
     private var player: Player
-    internal var drmPerformanceInfo: DrmPerformanceInfo?
     private var isStalling: Bool
     private var isSeeking: Bool
+    private var isMonitoring = false
+    
     /// DRM certificate download time in milliseconds
     private var drmCertificateDownloadTime: Int64?
-    private var isMonitoring = false
+    internal var drmDownloadTime: Int64?
+    private var drmType: String?
 
     init(player: Player, config: BitmovinAnalyticsConfig, stateMachine: StateMachine) {
         self.player = player
@@ -64,9 +66,8 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         }
 
         // drmType
-        if let drmType = self.drmPerformanceInfo?.drmType {
-            eventData.drmType = drmType
-        }
+        eventData.drmType = self.drmType
+        
 
         // videoBitrate
         if let bitrate = player.videoQuality?.bitrate {
@@ -120,10 +121,6 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         }
         player.remove(listener: self)
         isStalling = false
-    }
-    
-    func getDrmPerformanceInfo() -> DrmPerformanceInfo? {
-        return self.drmPerformanceInfo
     }
     
     var currentTime: CMTime? {
@@ -189,8 +186,8 @@ extension BitmovinPlayerAdapter: PlayerListener {
             // This request is the first that happens when initializing the DRM system
             self.drmCertificateDownloadTime = downloadTimeInMs
         case BMPHttpRequestTypeDrmLicenseFairplay:
-            let drmLoadTimeMs = (self.drmCertificateDownloadTime ?? 0) + (downloadTimeInMs ?? 0)
-            self.drmPerformanceInfo = DrmPerformanceInfo(drmType: DrmType.fairplay, drmLoadTime: drmLoadTimeMs)
+            self.drmDownloadTime = (self.drmCertificateDownloadTime ?? 0) + (downloadTimeInMs ?? 0)
+            self.drmType = DrmType.fairplay.rawValue
             self.drmCertificateDownloadTime = nil
         default:
             return