+++ BitmovinAnalyticsCollector/Classes/BitmovinPlayerV3/Classes/BitmovinAdAdapter.swift
@@ -1,5 +1,4 @@
 import BitmovinPlayer
-import Foundation
 
 public class BitmovinAdAdapter : NSObject, AdAdapter {
     private var bitmovinPlayer: Player
@@ -22,44 +21,44 @@ public class BitmovinAdAdapter : NSObject, AdAdapter {
     }
     
     func isAutoPlayEnabled() -> Bool {
-        self.bitmovinPlayer.config.playbackConfiguration.isAutoplayEnabled
+        self.bitmovinPlayer.config.playbackConfig.isAutoplayEnabled
     }
 }
 
 extension BitmovinAdAdapter : PlayerListener {
-    public func onAdManifestLoaded(_ event: AdManifestLoadedEvent) {
+    public func onAdManifestLoaded(_ event: AdManifestLoadedEvent, player: Player) {
         self.adAnalytics.onAdManifestLoaded(adBreak: AdModelMapper.fromPlayerAdConfiguration(adConfiguration: event.adBreak), downloadTime: event.downloadTime)
     }
     
-    public func onAdStarted(_ event: AdStartedEvent) {
+    public func onAdStarted(_ event: AdStartedEvent, player: Player) {
         self.adAnalytics.onAdStarted(ad: AdModelMapper.fromPlayerAd(playerAd: event.ad))
     }
     
-    public func onAdFinished(_ event: AdFinishedEvent) {
+    public func onAdFinished(_ event: AdFinishedEvent, player: Player) {
         self.adAnalytics.onAdFinished()
     }
     
-    public func onAdBreakStarted(_ event: AdBreakStartedEvent) {
+    public func onAdBreakStarted(_ event: AdBreakStartedEvent, player: Player) {
         self.adAnalytics.onAdBreakStarted(adBreak: AdModelMapper.fromPlayerAdConfiguration(adConfiguration: event.adBreak))
     }
     
-    public func onAdBreakFinished(_ event: AdBreakFinishedEvent) {
+    public func onAdBreakFinished(_ event: AdBreakFinishedEvent, player: Player) {
         self.adAnalytics.onAdBreakFinished()
     }
     
-    public func onAdClicked(_ event: AdClickedEvent) {
+    public func onAdClicked(_ event: AdClickedEvent, player: Player) {
         self.adAnalytics.onAdClicked(clickThroughUrl: event.clickThroughUrl?.absoluteString)
     }
     
-    public func onAdSkipped(_ event: AdSkippedEvent) {
+    public func onAdSkipped(_ event: AdSkippedEvent, player: Player) {
         self.adAnalytics.onAdSkipped()
     }
         
-    public func onAdError(_ event: AdErrorEvent) {
+    public func onAdError(_ event: AdErrorEvent, player: Player) {
         self.adAnalytics.onAdError(adBreak: AdModelMapper.fromPlayerAdConfiguration(adConfiguration: event.adConfig), code: Int(event.code), message: event.message)
     }
     
-    public func onAdQuartile(_ event: AdQuartileEvent) {
+    public func onAdQuartile(_ event: AdQuartileEvent, player: Player) {
         self.adAnalytics.onAdQuartile(quartile: BitmovinPlayerUtil.getAdQuartileFromPlayerAdQuartile(adQuartile: event.adQuartile))
     }
 }