+++ BitmovinAnalyticsCollector/Classes/BitmovinPlayerV3/Classes/BitmovinPlayerAdapter.swift
@@ -1,15 +1,15 @@
-import Foundation
 import BitmovinPlayer
 
 class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
     private let config: BitmovinAnalyticsConfig
     private var player: Player
-    internal var drmPerformanceInfo: DrmPerformanceInfo?
     private var isStalling: Bool
     private var isSeeking: Bool
+    private var isMonitoring = false
+    
     /// DRM certificate download time in milliseconds
     private var drmCertificateDownloadTime: Int64?
-    private var isMonitoring = false
+    internal var drmDownloadTime: Int64?
 
     init(player: Player, config: BitmovinAnalyticsConfig, stateMachine: StateMachine) {
         self.player = player
@@ -33,39 +33,58 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         //PlayerTech
         eventData.playerTech = "ios:bitmovin"
 
-        //Duration
-        if !player.duration.isNaN && !player.duration.isInfinite {
-            eventData.videoDuration = Int64(player.duration * BitmovinAnalyticsInternal.msInSec)
-        }
-
         //isCasting
         eventData.isCasting = player.isCasting
-
-        //isLive
-        eventData.isLive = self.isPlayerReady ? player.isLive : self.config.isLive
 
         //version
         if let sdkVersion = BitmovinPlayerUtil.playerVersion() {
             eventData.version = PlayerType.bitmovin.rawValue + "-" + sdkVersion
         }
         
-        let sourceUrl = player.config.sourceItem?.url(forType: player.streamType)
-        switch player.streamType {
-        case .DASH:
-            eventData.streamFormat = StreamType.dash.rawValue
-            eventData.mpdUrl = sourceUrl?.absoluteString
-        case .HLS:
-            eventData.streamFormat = StreamType.hls.rawValue
-            eventData.m3u8Url = sourceUrl?.absoluteString
-        case .progressive:
-            eventData.streamFormat = StreamType.progressive.rawValue
-            eventData.progUrl = sourceUrl?.absoluteString
-        default: break;
-        }
-
-        // drmType
-        if let drmType = self.drmPerformanceInfo?.drmType {
-            eventData.drmType = drmType
+        if let source = player.source {
+            let sourceConfig = source.sourceConfig
+            // streamFormat & urls
+            switch sourceConfig.type {
+                case SourceType.dash:
+                    eventData.streamFormat = StreamType.dash.rawValue
+                    eventData.mpdUrl = sourceConfig.url.absoluteString
+                case SourceType.hls:
+                    eventData.streamFormat = StreamType.hls.rawValue
+                    eventData.m3u8Url = sourceConfig.url.absoluteString
+                case SourceType.progressive:
+                    eventData.streamFormat = StreamType.progressive.rawValue
+                    eventData.progUrl = sourceConfig.url.absoluteString
+                default: break;
+            }
+            
+            // isLive & duration
+            let duration = source.duration
+            if (duration == 0) {
+                eventData.isLive = config.isLive
+            } else {
+                if (duration.isInfinite) {
+                    eventData.isLive = true;
+                } else {
+                    eventData.isLive = false;
+                    eventData.videoDuration = duration.milliseconds ?? 0
+                }
+            }
+            
+            // drmType
+            if let drmConfig = sourceConfig.drmConfig {
+                if (drmConfig is WidevineConfig) {
+                    eventData.drmType = DrmType.widevine.rawValue
+                } else if (drmConfig is PlayReadyConfig) {
+                    eventData.drmType = DrmType.playready.rawValue
+                } else if (drmConfig is FairplayConfig) {
+                    eventData.drmType = DrmType.fairplay.rawValue
+                } else if (drmConfig is ClearKeyConfig) {
+                    eventData.drmType = DrmType.clearkey.rawValue
+                }
+            }
+        } else {
+            // player active Source is not available
+            eventData.isLive = config.isLive
         }
 
         // videoBitrate
@@ -122,19 +141,23 @@ class BitmovinPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         isStalling = false
     }
     
-    func getDrmPerformanceInfo() -> DrmPerformanceInfo? {
-        return self.drmPerformanceInfo
-    }
-    
     var currentTime: CMTime? {
         get {
             return Util.timeIntervalToCMTime(_: player.currentTime)
         }
     }
+    
+    func onErrorEvent(_ errorData: ErrorData) {
+        if (!stateMachine.didStartPlayingVideo && stateMachine.didAttemptPlayingVideo) {
+            stateMachine.onPlayAttemptFailed(withError: errorData)
+        } else {
+            stateMachine.error(withError: errorData, time: Util.timeIntervalToCMTime(_: player.currentTime))
+        }
+    }
 }
 
 extension BitmovinPlayerAdapter: PlayerListener {
-    func onPlay(_ event: PlayEvent) {
+    func onPlay(_ event: PlayEvent, player: Player) {
         stateMachine.play(time: nil)
         
         if (isStalling && stateMachine.state != .seeking && stateMachine.state != .buffering) {
@@ -142,46 +165,46 @@ extension BitmovinPlayerAdapter: PlayerListener {
         }
     }
     
-    func onPlaying(_ event: PlayingEvent) {
+    func onPlaying(_ event: PlayingEvent, player: Player) {
         if (!isSeeking && !isStalling) {
             stateMachine.playing(time: Util.timeIntervalToCMTime(_: player.currentTime))
         }
     }
 
-    func onAdBreakStarted(_ event: AdBreakStartedEvent) {
+    func onAdBreakStarted(_ event: AdBreakStartedEvent, player: Player) {
         stateMachine.transitionState(destinationState: .ad, time: currentTime)
     }
     
-    func onAdBreakFinished(_ event: AdBreakFinishedEvent) {
+    func onAdBreakFinished(_ event: AdBreakFinishedEvent, player: Player) {
         stateMachine.transitionState(destinationState: .adFinished, time: currentTime)
     }
     
-    func onPaused(_ event: PausedEvent) {
+    func onPaused(_ event: PausedEvent, player: Player) {
         isSeeking = false
         stateMachine.pause(time: currentTime)
     }
 
-    func onReady(_ event: ReadyEvent) {
+    func onReady(_ event: ReadyEvent, player: Player) {
         self.isPlayerReady = true
     }
 
-    func onStallStarted(_ event: StallStartedEvent) {
+    func onStallStarted(_ event: StallStartedEvent, player: Player) {
         isStalling = true
         stateMachine.transitionState(destinationState: .buffering, time: Util.timeIntervalToCMTime(_: player.currentTime))
         
     }
 
-    func onStallEnded(_ event: StallEndedEvent) {
+    func onStallEnded(_ event: StallEndedEvent, player: Player) {
         isStalling = false
         transitionToPausedOrBufferingOrPlaying()
     }
 
-    func onSeek(_ event: SeekEvent) {
+    func onSeek(_ event: SeekEvent, player: Player) {
         isSeeking = true
         stateMachine.transitionState(destinationState: .seeking, time: Util.timeIntervalToCMTime(_: player.currentTime))
     }
 
-    func onDownloadFinished(_ event: DownloadFinishedEvent) {
+    func onDownloadFinished(_ event: DownloadFinishedEvent, player: Player) {
         let downloadTimeInMs = event.downloadTime.milliseconds
 
         switch event.downloadType {
@@ -189,8 +212,7 @@ extension BitmovinPlayerAdapter: PlayerListener {
             // This request is the first that happens when initializing the DRM system
             self.drmCertificateDownloadTime = downloadTimeInMs
         case BMPHttpRequestTypeDrmLicenseFairplay:
-            let drmLoadTimeMs = (self.drmCertificateDownloadTime ?? 0) + (downloadTimeInMs ?? 0)
-            self.drmPerformanceInfo = DrmPerformanceInfo(drmType: DrmType.fairplay, drmLoadTime: drmLoadTimeMs)
+            self.drmDownloadTime = (self.drmCertificateDownloadTime ?? 0) + (downloadTimeInMs ?? 0)
             self.drmCertificateDownloadTime = nil
         default:
             return
@@ -201,7 +223,7 @@ extension BitmovinPlayerAdapter: PlayerListener {
         return old?.bitrate != new?.bitrate
     }
 
-    func onVideoDownloadQualityChanged(_ event: VideoDownloadQualityChangedEvent) {
+    func onVideoDownloadQualityChanged(_ event: VideoDownloadQualityChangedEvent, player: Player) {
         let videoBitrateDidChange = didVideoBitrateChange(old: event.videoQualityOld, new: event.videoQualityNew)
         // there is a qualityChange event happening before the `onReady` method. Do not transition into any state.
         if isPlayerReady && !isStalling && !isSeeking && videoBitrateDidChange {
@@ -211,32 +233,28 @@ extension BitmovinPlayerAdapter: PlayerListener {
     }
     
     // No check if audioBitrate changes because no data available
-    func onAudioChanged(_ event: AudioChangedEvent) {
+    func onAudioChanged(_ event: AudioChangedEvent, player: Player) {
         if isPlayerReady && !isStalling && !isSeeking {
             stateMachine.audioQualityChange(time: currentTime)
             transitionToPausedOrBufferingOrPlaying()
         }
     }
 
-    func onSeeked(_ event: SeekedEvent) {
+    func onSeeked(_ event: SeekedEvent, player: Player) {
         isSeeking = false
         if (!isStalling) {
             transitionToPausedOrBufferingOrPlaying()
         }
     }
 
-    func onPlaybackFinished(_ event: PlaybackFinishedEvent) {
+    func onPlaybackFinished(_ event: PlaybackFinishedEvent, player: Player) {
         stateMachine.transitionState(destinationState: .paused, time: Util.timeIntervalToCMTime(_: player.duration))
         stateMachine.disableHeartbeat()
     }
 
-    func onError(_ event: ErrorEvent) {
-        let errorData = ErrorData(code: Int(event.code), message: event.message, data: nil)
-        if (!stateMachine.didStartPlayingVideo && stateMachine.didAttemptPlayingVideo) {
-            stateMachine.onPlayAttemptFailed(withError: errorData)
-        } else {
-            stateMachine.error(withError: errorData, time: Util.timeIntervalToCMTime(_: player.currentTime))
-        }
+    func onPlayerError(_ event: PlayerErrorEvent, player: Player) {
+        let errorData = ErrorData(code: Int(event.code.rawValue), message: event.message, data: nil)
+        onErrorEvent(errorData)
     }
 
     func transitionToPausedOrBufferingOrPlaying() {
@@ -254,17 +272,17 @@ extension BitmovinPlayerAdapter: PlayerListener {
         }
     }
     
-    func onSourceWillUnload(_ event: SourceWillUnloadEvent) {
+    func onSourceUnload(_ event: SourceUnloadEvent, player: Player) {
         if (!stateMachine.didStartPlayingVideo && stateMachine.didAttemptPlayingVideo) {
             stateMachine.onPlayAttemptFailed(withReason: VideoStartFailedReason.pageClosed)
         }
     }
     
-    func onSourceUnloaded(_ event: SourceUnloadedEvent) {
+    func onSourceUnloaded(_ event: SourceUnloadedEvent, player: Player) {
         stateMachine.reset()
     }
     
-    func onSubtitleChanged(_ event: SubtitleChangedEvent) {
+    func onSubtitleChanged(_ event: SubtitleChangedEvent, player: Player) {
         guard stateMachine.state == .paused || stateMachine.state == .playing else {
             return
         }
@@ -272,4 +290,11 @@ extension BitmovinPlayerAdapter: PlayerListener {
         transitionToPausedOrBufferingOrPlaying()
     }
 
+}
+
+extension BitmovinPlayerAdapter: SourceListener {
+    func onSourceError(_ event: SourceErrorEvent, player: Player) {
+        let errorData = ErrorData(code: Int(event.code.rawValue), message: event.message, data: nil)
+        onErrorEvent(errorData)
+    }
 }