+++ BitmovinAnalyticsCollector/Classes/Collector/ads/AnalyticsAdTagType.swift
@@ -1,6 +1,4 @@
-import Foundation
-
-public enum AdTagType: String {
+public enum AnalyticsAdTagType: String {
     case VAST = "vast"
     case VMAP = "vmap"
     case VPAID = "vpaid"