+++ BitmovinAnalyticsCollector/Classes/Collector/enums/DrmType.swift
@@ -1,5 +1,6 @@
-import Foundation
-
-public struct DrmType {
-    public static let fairplay = "fairplay"
+public enum DrmType: String {
+    case fairplay = "fairplay"
+    case widevine = "widevine"
+    case playready = "playready"
+    case clearkey = "clearkey"
 }